pyproject.toml
[project]
name = "redshift-mcp-server"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for Amazon Redshift"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Tuan Nguyen"},
]
maintainers = [
    {name = "Tuan Nguyen"},
]
keywords = ["redshift", "mcp", "llm", "database"]
license = {text = "MIT"}
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mcp>=0.4.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
]

[project.scripts]
"redshift-mcp-server" = "redshift_mcp_server.server:main"

[project.urls]
Source = "https://github.com/tuanknguyen/redshift-mcp-server"
"Bug Tracker" = "https://github.com/tuanknguyen/redshift-mcp-server/issues"

[project.optional-dependencies]
dev = [
    "uvicorn>=0.23.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "ruff>=0.9.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["redshift_mcp_server"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that use real database connections (deselect with '-m \"not integration\"')",
    "asyncio: marks tests that use asyncio"
]
asyncio_mode = "strict"

[tool.ruff]
line-length = 99
exclude = [
    ".venv",
    "**/__pycache__",
    "**/dist",
    "**/build",
    "**/.ruff_cache",
]
force-exclude = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.uv]
cache-dir = ".uv/packages"